Don't auto-detect virtual axes
Using MFLAGS.#DEFCON is not a valid way to detect whether an axis is
virtual.  In fact, there's no reliable way to detect whether an axis
is virtual.  So, remove the auto-detection logic, and add support
for specifying a comma-separated list of virtual axes as a parameter
to the AcsMotionConfig IOC shell command and the constructor of the
SPiiPlusController class.